测试最佳执行线程数

+++ src/main/java/com/sunducation/waterflow/service/ImportDataService.java
@@ -6,5 +6,5 @@ package com.sunducation.waterflow.service;
  */
 public interface ImportDataService {
 
-  Long parseAndinsertDb(String fiepath) throws Exception;
+  double parseAndinsertDb(String fiepath) throws Exception;
 }
